+++ aquadoggo/src/graphql/mod.rs
@@ -2,10 +2,12 @@
 
 pub mod constants;
 pub mod mutations;
+pub mod queries;
 pub mod scalars;
 mod schema;
-mod schema_builders;
 pub mod types;
 pub mod utils;
+#[cfg(test)]
+mod tests;
 
 pub use schema::{build_root_schema, GraphQLSchemaManager};
