--- aquadoggo/src/graphql/schema_builders/mod.rs
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod document;
-mod document_field;
-mod next_args;
-#[cfg(test)]
-mod tests;
-
-pub use document::{build_all_document_query, build_document_query, build_document_schema};
-pub use document_field::build_document_field_schema;
-pub use next_args::build_next_args_query;
