+++ aquadoggo/src/graphql/types/mod.rs
@@ -2,6 +2,10 @@
 
 mod document_meta;
 mod next_arguments;
+mod document;
+mod document_fields;
 
 pub use document_meta::DocumentMeta;
 pub use next_arguments::NextArguments;
+pub use document::Document;
+pub use document_fields::DocumentFields;
